requirements.txt
astropy==4.1
astropy-healpix==0.5
attrs==20.3.0
cffi==1.14.4
corner==2.2.1
cycler==0.10.0
easyquery==0.1.6
emcee==3.0.2
future==0.18.2
healpy==1.14.0
importlib-metadata==3.7.3
iniconfig==1.1.1
kiwisolver==1.3.1
llvmlite==0.35.0
LSSTDESC.Coord==1.2.1
matplotlib==3.3.3
numba==0.52.0
numexpr==2.7.2
numpy==1.19.4
packaging==20.9
pandas==1.1.5
photutils==1.0.1
Pillow==8.1.0
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
regions==0.4
reproject==0.7.1
scipy==1.5.4
six==1.15.0
toml==0.10.2
tqdm==4.56.0
TreeCorr==4.1.5
typing-extensions==3.7.4.3
zipp==3.4.1
